+++ include/mruby_integration/text.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_text(mrb_state*);