+++ mrb_doc/text.rb
@@ -1,9 +1,0 @@
-# Draws the current frame rate at the passed in x and y coordinates.
-# @param x [Integer]
-# @param y [Integer]
-# @return [nil]
-def draw_fps(x, y)
-  # mrb_draw_fps
-  # src/mruby_integration/text.cpp
-  nil
-end